Split workbooks and views methods
